Fix JK flipflop logic
The D, SR, and T flip flops implemented currently don't worry about the enable time of the gate signal, so the JK shouldn't either.  It's ideal operation is with the rising trigger (as with D, SR and T).
As existing javascript code only changes outputs on rising edge, JK should match.